fix one of the streaming endpoints
